Update message views routes for PEP8 Compliance.

+++ wyspa/messages/views.py
@@ -38,32 +38,30 @@ def my_voice():
 @ login_required
 def create_wyspa():
 
-    try:
-        # Convert datetime
-        formatted_expiry = Wyspa.string_to_datetime(
-                expiry_date=request.form.get("expiryDate"),
-                expiry_time=request.form.get("expiryTime"))
-    except TypeError as e:
-        print(e)
-        flash("Oops! Something went wrong! Please try again!")
-        return redirect(url_for("messages.my_voice"))
-
-    # If the formatted_expiry fails
-    if not formatted_expiry:
-        flash("Expiry must be in the future!")
-        return redirect(url_for("messages.my_voice"))
-
-    # Try and convert the address to latlong
-    try:
-        converted_location = Wyspa.location_to_latlong(
-            request.form.get("location"))
-    except Exception as e:
-        print(e)
-        flash("Unable to locate address!")
-        return redirect(url_for("messages.my_voice"))
-
-    # Create a new wyspa, and write to DB
     if request.method == "POST":
+        try:
+            # Convert datetime
+            formatted_expiry = Wyspa.string_to_datetime(
+                    expiry_date=request.form.get("expiryDate"),
+                    expiry_time=request.form.get("expiryTime"))
+        except TypeError:
+            flash("Oops! Something went wrong! Please try again!")
+            return redirect(url_for("messages.my_voice"))
+
+        # If the formatted_expiry fails
+        if not formatted_expiry:
+            flash("Expiry must be in the future!")
+            return redirect(url_for("messages.my_voice"))
+
+        # Try and convert the address to latlong
+        try:
+            converted_location = Wyspa.location_to_latlong(
+                request.form.get("location"))
+        except AttributeError:
+            flash("Unable to locate address!")
+            return redirect(url_for("messages.my_voice"))
+
+        # Create a new wyspa, and write to DB
         new_wyspa = Wyspa(author=current_user.username,
                           message=request.form.get("wyspaContent"),
                           mood=int(request.form.get("mood")),
@@ -73,24 +71,29 @@ def create_wyspa():
         flash("Wyspa successful!")
         return redirect(url_for("messages.my_voice"))
 
+    # Get Route
+    return redirect(url_for("messages.my_voice"))
+
 
 # Add comment to existing Wyspa
 @ messages.route('/add_comment/<message_id>', methods=["GET", "POST"])
 @ login_required
 def add_comment(message_id):
 
-    if request.method == "POST":
+    # Retrieve Wyspa
+    retrieved_wyspa = Wyspa.get_by_id(message_id)
 
-        # Retrieve Wyspa
-        retrieved_wyspa = Wyspa.get_by_id(message_id)
+    # If comment posted
+    if request.method == "POST":
 
         # Check if user is logged in to add to comment db
         if current_user.is_authenticated:
             retrieved_wyspa.add_comment(request.form.get(
                 "commentReply"), current_user.username)
 
-        return redirect(url_for("messages.view_message",
-                                message_id=retrieved_wyspa._id))
+    # Both GET and POST requests
+    return redirect(url_for("messages.view_message",
+                            message_id=retrieved_wyspa.wyspa_id))
 
 
 # Remove comment from existing Wyspa
@@ -99,11 +102,19 @@ def add_comment(message_id):
 @ login_required
 def remove_comment(message_id, comment_id):
 
+    # Query DB for Wyspa
     retrieved_wyspa = Wyspa.get_by_id(message_id)
-    retrieved_wyspa.remove_comment(int(comment_id)-1)
 
+    if request.method == "POST":
+        # Ensure author is logged in user
+        if current_user.username == retrieved_wyspa.author:
+            retrieved_wyspa.remove_comment(int(comment_id)-1)
+        else:
+            flash("You are not authorised to delete this comment!")
+
+    # Get and Post Routes
     return redirect(url_for("messages.view_message",
-                            message_id=retrieved_wyspa._id))
+                            message_id=retrieved_wyspa.wyspa_id))
 
 
 # Add "Listen" to existing Wyspa
@@ -127,10 +138,10 @@ def add_listen(message_id, listener):
 
     # Inform the user they need to be logged in
     else:
-        flash("Sorry - you need to be logged in to listen to a Wyspa!")
+        flash("You must be logged in to listen to a Wyspa!")
 
     return redirect(url_for("messages.view_message",
-                            message_id=retrieved_wyspa._id))
+                            message_id=retrieved_wyspa.wyspa_id))
 
 
 # Delete WYSPA
@@ -139,8 +150,14 @@ def add_listen(message_id, listener):
 def remove_wyspa(message_id):
 
     if request.method == "POST":
-        Wyspa.delete_wyspa(message_id)
-        return redirect(url_for("messages.my_voice"))
+        # Retrieve the wyspa
+        retrieved_wyspa = Wyspa.get_by_id(message_id)
+        # Ensure logged in user is owner of Wyspa
+        if current_user.username == retrieved_wyspa.author:
+            Wyspa.delete_wyspa(message_id)
+
+    # Get and Post Routes
+    return redirect(url_for("messages.my_voice"))
 
 
 # Edit WYSPA
@@ -159,7 +176,7 @@ def edit_wyspa(message_id):
         flash("You can't edit someone elses Wyspa!")
         return redirect(url_for("messages.my_voice"))
 
-        # Post View
+    # Post Route
     if request.method == "POST":
 
         # Convert datetime
@@ -180,8 +197,7 @@ def edit_wyspa(message_id):
         try:
             converted_location = Wyspa.location_to_latlong(
                 request.form.get("location"))
-        except Exception as e:
-            print(e)
+        except AttributeError:
             flash("Unable to locate address!")
             expiry_date = Wyspa.datetime_to_string(
                 retrieved_wyspa.expiry)
@@ -198,8 +214,7 @@ def edit_wyspa(message_id):
         flash("Wyspa Updated!")
         return redirect(url_for("messages.my_voice"))
 
-    # Get view
-
+    # Get Route
     expiry_date = Wyspa.datetime_to_string(
         retrieved_wyspa.expiry)
 
@@ -220,10 +235,10 @@ def delete_wyspa(message_id):
         if not retrieved_wyspa.author == current_user.username:
             flash("You do not have access to this Wyspa!")
             return redirect(url_for("messages.my_voice"))
+
         # Delete Wyspa
-        else:
-            Wyspa.delete_wyspa(message_id)
-            return redirect(url_for("messages.my_voice"))
+        Wyspa.delete_wyspa(message_id)
+        return redirect(url_for("messages.my_voice"))
 
     # Route for GET
     return redirect(url_for("messages.my_voice"))
